generalize the modal form
