+++ utils/item-display-utils.ts
@@ -1,0 +1,6 @@
+export const getDate = (iso: string) => {
+    const date = new Date(iso);
+    return `${date.toLocaleDateString("en-US")} ${date.toLocaleTimeString(
+      "en-US"
+    )}`;
+  };
